+++ docs/Changelog.md
@@ -9,4 +9,5 @@
   - Added JSON path assertion.
   - Added parameters for final_stats reporter
   - Added ability to generate query string based on parameters and request type.
+  - Added pass/fail criteria report on console reporter.